+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/entity/oddion/OddionModel.java
@@ -3,7 +3,6 @@ package net.potionstudios.biomeswevegone.world.entity.oddion;
 import net.minecraft.resources.ResourceLocation;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import software.bernie.geckolib.core.animatable.GeoAnimatable;
-import software.bernie.geckolib.core.animation.Animation;
 import software.bernie.geckolib.model.GeoModel;
 
 /**
@@ -11,7 +10,7 @@ import software.bernie.geckolib.model.GeoModel;
  * @see GeoModel
  * @author YaBoiChips, Joseph T. McQuigg
  */
-public class OddionModel<T extends GeoAnimatable> extends GeoModel<T> {
+class OddionModel<T extends GeoAnimatable> extends GeoModel<T> {
 
     @Override
     public ResourceLocation getModelResource(T object) {
@@ -26,10 +25,5 @@ public class OddionModel<T extends GeoAnimatable> extends GeoModel<T> {
     @Override
     public ResourceLocation getAnimationResource(T animatable) {
         return BiomesWeveGone.id("animations/oddion.animation.json");
-    }
-
-    @Override
-    public Animation getAnimation(T animatable, String name) {
-        return super.getAnimation(animatable, name);
     }
 }